Update: Remove old packages

+++ all_in.sh
@@ -20,15 +20,6 @@ db_name="fermiarcs"
 rm ${repo_path}/*.sig
 rm ${repo_path}/${db_name}.*
 
-# for i in ${all_pkg}; do
-
-#   # sign all
-#   gpg --batch --passphrase-file your_password_file --pinentry-mode loopback --detach-sign "${repo_path}/${i}"
-
-#   # build repo db
-#   repo-add "${repo_path}/${db_name}.db.tar.gz" "${repo_path}/${i}"
-# done
-
 for i in ${all_pkg}; do
   # get the package name
   package=${i}
@@ -37,35 +28,33 @@ for i in ${all_pkg}; do
     package=$(echo ${package} | sed "s/-${remove}//")
   done
 
+  # get the version name
   if [ "$(find ${repo_path} -name "${package}-[0-9]*.pkg.tar.xz" | wc -l)" -gt "1" ]; then
     remove=$(echo ${i} | awk -F '-' '{print $NF}' | head -n 1)
     version=$(echo ${i} | sed "s/-${remove}//g" | sed "s/${package}-//g" | awk -F ' ' '{print $1}')
     # version: 69.0b6-1
 
+    # delete the old version
     if [[ -f ${repo_path}/${package}_version.txt ]]; then
       echo "${package}_version.txt exist"
-      # version_before=$(cat ${package}_version.txt)
-      # echo ${version} | tee -a ${package}_version.txt
-      # bigger=$(sort -rV ver_test.txt | head -n 1)
-      # echo "bigger: ${bigger}"
-      # if [[ ${bigger} > ${version_before} ]]; then
-      gpg --batch --passphrase-file your_password_file --pinentry-mode loopback --detach-sign "${repo_path}/${i}"
-      #   repo-remove "${repo_path}/${db_name}.db.tar.gz" "${repo_path}/${package}"
-      #   repo-add "${repo_path}/${db_name}.db.tar.gz" "${repo_path}/${i}"
-      # fi
+      rm ${repo_path}/${i}*
 
+    # save the newest version
     else
-
       echo ${version} >${repo_path}/${package}_version.txt
       gpg --batch --passphrase-file your_password_file --pinentry-mode loopback --detach-sign "${repo_path}/${i}"
       repo-add "${repo_path}/${db_name}.db.tar.gz" "${repo_path}/${i}"
 
     fi
-
+  # only have one version
   else
-    gpg --batch --passphrase-fileyour_password_file --pinentry-mode loopback --detach-sign "${repo_path}/${i}"
+    gpg --batch --passphrase-file your_password_file --pinentry-mode loopback --detach-sign "${repo_path}/${i}"
     repo-add "${repo_path}/${db_name}.db.tar.gz" "${repo_path}/${i}"
 
   fi
 done
-rm ${repo_path}/*_version.txt
+
+# delete ${repo_path}/*_version.txt
+if [ -n "$(find ${repo_path} -maxdepth 1 -name '*_version.txt')" ]; then
+  rm ${repo_path}/*_version.txt
+fi
